Update to consider characters as strings of utf8 codeoints in most places

+++ src/c_config.h
@@ -14,7 +14,7 @@
 #include "sysdep.h"
 
 typedef struct _CurPos {
-    unsigned int sz;
+    size_t sz;
     char *a;
     char *c;
     char *z;
